GUI/requirements.txt
Flask==2.3.0
mido==1.3.0
pyserial==3.5